pyproject.toml
[project]
name = "agents-monitor-schema"
version = "1.0.2"
description = "Civic Interconnect Schema Monitoring Agent"
authors = [
    { name="Civic Interconnect", email="info@civic-interconnect.org" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "python-dotenv",
    "PyYAML",
    "typer[all]",
    "rich",
    "GitPython",
    "requests",
    "gql[aiohttp]",
    "pandas",
    "civic-lib-core @ git+https://github.com/civic-interconnect/civic-lib-core.git@main"
]

[project.optional-dependencies]
dev = [
  "mkdocs",
  "mkdocs-material",
  "pre-commit",
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "ruff",
]

[project.urls]
"Homepage" = "https://github.com/civic-interconnect/civic-lib-core"
"Source" = "https://github.com/civic-interconnect/civic-lib-core"

[project.scripts]
civic-dev = "civic_dev.cli:app"
schema-agent = "agents_monitor_schema.cli.cli:app"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
fix = true
unsafe-fixes = true
preview = true

[tool.ruff.lint]
extend-select = ["E", "F", "W", "I", "UP", "C90", "N", "B", "A", "SIM"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.format]
docstring-code-format = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra --strict-markers"
testpaths = [
  "tests"
]
python_files = "test_*.py"
